Misc PR adjustments; refactor + reorg; unit test WIP.
